requirements.txt
pyyaml
textual
textual-terminal
rich-pixels
pyalsaaudio
